Move collection forms to wagtail.admin.forms.collections
